Add Ability to Add a Battleship on Click
Still a work in progress.

Some unwanted interactions: clicking on the enemies board adds the ship to the player's board. I want the user to place shits from decending order. so after a ship is placed, a different ship should be placed, and so on and so forth. Currently I can only add ships vertically.