CPU Attacks on Player Board is Now Show
Need to fix the randomAttack function. Not fully working as intended. if it chooses an occupied spot, it is not rerunning to hit an empty spot instead. Maybe I should have an array of unhit spots that the cpu will always attack. this way it wont have to ever rerun if it selects a hit location. it will  randomly chose out of a list of unhit locations everytime