Partially Working cpu-logic.js
The CPU will go in the direction it chooses after it hits a ship. I need to fix a problem where it keeps going in that direction when it misses. for now i want it to choose a random spot after is misses. but my end goal is to have the cpu go in the opposite direction after a miss, until it sinks the ship it is currently attacking