Add a backTracker
this is for when the cpu misses after hitting a ship. it will track what coord was a hit before that miss. this is based on the opposite direction the cpu was traveling before it missed. this will allow me to backtrack continously untl the it destroys the ship it is currently on